src/Display/Bitmaps/Scripts/requirements.txt
pypng>=0.0.20